dematerialize CommonMirror into extension methods in semantic.Api
